Added a version call.
The idea is that we never call communicate() to ask for the version. Because
the multiplus sends version frames when it gets bored, bad timing means we
might receive a version frame when asking for something else. This way, we
can safely discard unexpected version frames. Note: We are not doing such
discarding yet.